Consistently use type float for result in OutputTree

+++ twogtp/OutputTree.cpp
@@ -25,11 +25,11 @@ using libpentobi_base::get_transformed;
 namespace {
 
 void add(PentobiTree& tree, const SgfNode& node, bool is_player_black,
-         bool is_real_move, double result)
+         bool is_real_move, float result)
 {
     unsigned index = is_player_black ? 0 : 1;
     array<unsigned, 2> count;
-    array<double, 2> avg_result;
+    array<float, 2> avg_result;
     array<unsigned, 2> real_count;
     auto comment = tree.get_comment(node);
     if (comment.empty())
@@ -104,7 +104,7 @@ OutputTree::OutputTree(Variant variant)
 OutputTree::~OutputTree() = default; // Non-inline to avoid GCC -Winline warning
 
 void OutputTree::add_game(const Board& bd, unsigned player_black,
-                          double result, const array<bool,
+                          float result, const array<bool,
                           Board::max_moves>& is_real_move)
 {
     if (bd.has_setup())
